+++ docs/changelog.rst
@@ -1,6 +1,12 @@
 Changelog
 ---------
 
+v0.9.4
+^^^^^^
+
+* Fixed error on invalid chapter redirect
+* Fixed external redirects being allowed
+
 v0.9.3
 ^^^^^^